requirements.txt
python-telegram-bot==13.4.1
toml==0.10.2
pillow==8.4.0